ci: Use Meson version from pacman on msys2 CI runners
Rather than pinning it to the lowest version we support, as is the
standard policy.

This means we’ll end up using version 1.3.2-2, which has just been
packaged to contain the fix for
https://github.com/mesonbuild/meson/issues/12330, which has been
impacting GLib significantly since we started installing
gobject-introspection in CI in commit
c428d6e67318e7b02c13111db260ed98622e982a.

Thanks to Christoph Reiter, Luca Bacci and Simon McVittie for diagnosing
and fixing this issue.

Signed-off-by: Philip Withnall <pwithnall@gnome.org>

Fixes: #3262