meson: fallback to proxy-libintl if intl is not found by meson
We require this to build in platforms in which intl is not provided
(like android) by the system.

Meson does not allow to use pkg-config to resolve 'intl' as it has a
specific builtin method to do so, so we provide it via the 'proxy'.

Also avoid 'intl' dependency on Apple platforms

It causes linking problems.